[Transaction] Fix initTransaction might wait until request timeout (#1991)
### Motivation

Most of the tests in `TransactionTest` take at least 10 seconds because
the `request.timeout.ms` config is 10000. See

https://github.com/streamnative/kop/blob/1fd3bdb9158fd944c2fb7a241c0d8dca923367ab/tests/src/test/java/io/streamnative/pulsar/handlers/kop/coordinator/transaction/TransactionTest.java#L1150

It's because KoP loads the metadata topic lazily, unlike Kafka, which
loads the metadata topic when started. KoP adopts this way because KoP
starts before the Pulsar broker finishes the start, in this case there
might be some problems with the topic lookup.

Therefore, when the Kafka producer sends the 1st INIT_PRODUCER_ID
request, it will receive `COORDINATOR_LOAD_IN_PROGRESS` error, and the
request will reenqueue [1], then the request will expire after
`request.timeout.ms` milliseconds.

```java
} else if (error != Errors.NOT_COORDINATOR && error != Errors.COORDINATOR_NOT_AVAILABLE) {
    if (error != Errors.COORDINATOR_LOAD_IN_PROGRESS && error != Errors.CONCURRENT_TRANSACTIONS) {
       /* ... */
    } else {
        this.reenqueue(); // [1]
    }
} else { // [2]
    TransactionManager.this.lookupCoordinator(CoordinatorType.TRANSACTION, TransactionManager.this.transactionalId);
    this.reenqueue();
}
```

### Modifications

Return `COORDINATOR_NOT_AVAILABLE` instead of
`COORDINATOR_LOAD_IN_PROGRESS` when the transaction metadata topic is
loading. It's different with Kafka's behavior but it can avoid waiting
30 seconds by default for `initTransaction` in KoP.

In addition, configure the request timeout with 3 seconds for some cases
that `initTransaction` might expire so that the test time could be
reduced.