[fix] Memory leak of the map in ThreadLocalAccessor
Fix: https://github.com/streamnative/kop/issues/1956

### Modifications
Update stats on the netty threads to fix the stats memory leak issue.

The root cause is we can only use `DataSketchesOpStatsLogger` in the netty thread.
When used in other threads, it will cause memory leak, since the `FastThreadLocal` can't remove the map if the thread is not netty thread.